8200298: Unify all unix versions of libjsig/jsig.c Reviewed-by: dholmes, stuefe
