+++ make/mapfiles/libjsig/mapfile-vers-solaris
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2005, 2013, Oracle and/or its affiliates. All rights reserved.
+# Copyright (c) 2005, 2018, Oracle and/or its affiliates. All rights reserved.
 # DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
 #
 # This code is free software; you can redistribute it and/or modify it
@@ -28,7 +28,6 @@ SUNWprivate_1.1 {
         global:
             JVM_begin_signal_setting;
             JVM_end_signal_setting;
-            JVM_get_libjsig_version;
             JVM_get_signal_action;
             sigaction;
             signal;
