+++ src/java.base/unix/native/libjsig/jsig.c
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2001, 2015, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2001, 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2015 SAP SE. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,50 +23,80 @@
  *
  */
 
-/* CopyrightVersion 1.2 */
-
 /* This is a special library that should be loaded before libc &
  * libthread to interpose the signal handler installation functions:
  * sigaction(), signal(), sigset().
  * Used for signal-chaining. See RFE 4381843.
  */
 
-#include <signal.h>
 #include <dlfcn.h>
+#include <errno.h>
 #include <pthread.h>
+#include <signal.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <stdbool.h>
-#include <stdint.h>
+#include <string.h>
 
-#define MASK(sig) ((uint32_t)1 << (sig-1))  // 0 is not a signal.
-#if (32 < NSIG-1)
-#error "Not all signals can be encoded in jvmsigs. Adapt its type!"
+#if (__STDC_VERSION__ >= 199901L)
+  #include <stdbool.h>
+#else
+  #define bool int
+  #define true 1
+  #define false 0
 #endif
-static struct sigaction sact[NSIG]; /* saved signal handlers */
-static uint32_t jvmsigs = 0; /* signals used by jvm */
-static __thread bool reentry = false; /* prevent reentry deadlock (per-thread) */
 
-/* used to synchronize the installation of signal handlers */
+#ifdef SOLARIS
+#define MAX_SIGNALS (SIGRTMAX+1)
+
+/* On solaris, MAX_SIGNALS is a macro, not a constant, so we must allocate sact dynamically. */
+static struct sigaction *sact = (struct sigaction *)NULL; /* saved signal handlers */
+#else
+#define MAX_SIGNALS NSIG
+
+static struct sigaction sact[MAX_SIGNALS]; /* saved signal handlers */
+#endif
+
+static sigset_t jvmsigs; /* Signals used by jvm. */
+
+#ifdef MACOSX
+static __thread bool reentry = false; /* prevent reentry deadlock (per-thread) */
+#endif
+
+/* Used to synchronize the installation of signal handlers. */
 static pthread_mutex_t mutex = PTHREAD_MUTEX_INITIALIZER;
 static pthread_cond_t cond = PTHREAD_COND_INITIALIZER;
 static pthread_t tid = 0;
 
 typedef void (*sa_handler_t)(int);
 typedef void (*sa_sigaction_t)(int, siginfo_t *, void *);
-typedef sa_handler_t (*signal_t)(int, sa_handler_t);
+typedef sa_handler_t (*signal_function_t)(int, sa_handler_t);
 typedef int (*sigaction_t)(int, const struct sigaction *, struct sigaction *);
 
-static signal_t os_signal = 0; /* os's version of signal()/sigset() */
+static signal_function_t os_signal = 0; /* os's version of signal()/sigset() */
 static sigaction_t os_sigaction = 0; /* os's version of sigaction() */
 
 static bool jvm_signal_installing = false;
 static bool jvm_signal_installed = false;
 
+
+/* assume called within signal_lock */
+static void allocate_sact() {
+#ifdef SOLARIS
+  if (sact == NULL) {
+    sact = (struct sigaction *)malloc((MAX_SIGNALS) * (size_t)sizeof(struct sigaction));
+    if (sact == NULL) {
+      printf("%s\n", "libjsig.so unable to allocate memory");
+      exit(0);
+    }
+    memset(sact, 0, (MAX_SIGNALS) * (size_t)sizeof(struct sigaction));
+  }
+#endif
+}
+
 static void signal_lock() {
   pthread_mutex_lock(&mutex);
   /* When the jvm is installing its set of signal handlers, threads
-   * other than the jvm thread should wait */
+   * other than the jvm thread should wait. */
   if (jvm_signal_installing) {
     if (tid != pthread_self()) {
       pthread_cond_wait(&cond, &mutex);
@@ -83,41 +114,75 @@ static sa_handler_t call_os_signal(int sig, sa_handler_t disp,
 
   if (os_signal == NULL) {
     if (!is_sigset) {
-      os_signal = (signal_t)dlsym(RTLD_NEXT, "signal");
+      os_signal = (signal_function_t)dlsym(RTLD_NEXT, "signal");
     } else {
-      os_signal = (signal_t)dlsym(RTLD_NEXT, "sigset");
+      os_signal = (signal_function_t)dlsym(RTLD_NEXT, "sigset");
     }
     if (os_signal == NULL) {
       printf("%s\n", dlerror());
       exit(0);
     }
   }
+
+#ifdef MACOSX
+  /* On macosx, the OS implementation of signal calls sigaction.
+   * Make sure we do not deadlock with ourself. (See JDK-8072147). */
   reentry = true;
+#endif
+
   res = (*os_signal)(sig, disp);
+
+#ifdef MACOSX
   reentry = false;
+#endif
+
   return res;
 }
 
-static void save_signal_handler(int sig, sa_handler_t disp) {
+static void save_signal_handler(int sig, sa_handler_t disp, bool is_sigset) {
   sigset_t set;
+
   sact[sig].sa_handler = disp;
   sigemptyset(&set);
   sact[sig].sa_mask = set;
-  sact[sig].sa_flags = 0;
+  if (!is_sigset) {
+#ifdef SOLARIS
+    sact[sig].sa_flags = SA_NODEFER;
+    if (sig != SIGILL && sig != SIGTRAP && sig != SIGPWR) {
+      sact[sig].sa_flags |= SA_RESETHAND;
+    }
+#else
+    sact[sig].sa_flags = 0;
+#endif
+  } else {
+    sact[sig].sa_flags = 0;
+  }
 }
 
 static sa_handler_t set_signal(int sig, sa_handler_t disp, bool is_sigset) {
   sa_handler_t oldhandler;
   bool sigused;
+  bool sigblocked;
 
   signal_lock();
+  allocate_sact();
 
-  sigused = (MASK(sig) & jvmsigs) != 0;
+  sigused = sigismember(&jvmsigs, sig);
   if (jvm_signal_installed && sigused) {
     /* jvm has installed its signal handler for this signal. */
     /* Save the handler. Don't really install it. */
+    if (is_sigset) {
+      sigblocked = sigismember(&(sact[sig].sa_mask), sig);
+    }
     oldhandler = sact[sig].sa_handler;
-    save_signal_handler(sig, disp);
+    save_signal_handler(sig, disp, is_sigset);
+
+#ifdef SOLARIS
+    if (is_sigset && sigblocked) {
+      /* We won't honor the SIG_HOLD request to change the signal mask */
+      oldhandler = SIG_HOLD;
+    }
+#endif
 
     signal_unlock();
     return oldhandler;
@@ -126,10 +191,10 @@ static sa_handler_t set_signal(int sig, sa_handler_t disp, bool is_sigset) {
      * handlers and save the old ones. jvm uses sigaction().
      * Leave the piece here just in case. */
     oldhandler = call_os_signal(sig, disp, is_sigset);
-    save_signal_handler(sig, oldhandler);
+    save_signal_handler(sig, oldhandler, is_sigset);
 
     /* Record the signals used by jvm */
-    jvmsigs |= MASK(sig);
+    sigaddset(&jvmsigs, sig);
 
     signal_unlock();
     return oldhandler;
@@ -144,13 +209,27 @@ static sa_handler_t set_signal(int sig, sa_handler_t disp, bool is_sigset) {
 }
 
 sa_handler_t signal(int sig, sa_handler_t disp) {
+  if (sig < 0 || sig >= MAX_SIGNALS) {
+    errno = EINVAL;
+    return SIG_ERR;
+  }
+
   return set_signal(sig, disp, false);
 }
 
 sa_handler_t sigset(int sig, sa_handler_t disp) {
+#ifdef _ALLBSD_SOURCE
   printf("sigset() is not supported by BSD");
   exit(0);
- }
+#else
+  if (sig < 0 || sig >= MAX_SIGNALS) {
+    errno = EINVAL;
+    return (sa_handler_t)-1;
+  }
+
+  return set_signal(sig, disp, true);
+#endif
+}
 
 static int call_os_sigaction(int sig, const struct sigaction  *act,
                              struct sigaction *oact) {
@@ -169,13 +248,21 @@ int sigaction(int sig, const struct sigaction *act, struct sigaction *oact) {
   bool sigused;
   struct sigaction oldAct;
 
+  if (sig < 0 || sig >= MAX_SIGNALS) {
+    errno = EINVAL;
+    return -1;
+  }
+
+#ifdef MACOSX
   if (reentry) {
     return call_os_sigaction(sig, act, oact);
   }
+#endif
 
   signal_lock();
 
-  sigused = (MASK(sig) & jvmsigs) != 0;
+  allocate_sact();
+  sigused = sigismember(&jvmsigs, sig);
   if (jvm_signal_installed && sigused) {
     /* jvm has installed its signal handler for this signal. */
     /* Save the handler. Don't really install it. */
@@ -197,8 +284,8 @@ int sigaction(int sig, const struct sigaction *act, struct sigaction *oact) {
       *oact = oldAct;
     }
 
-    /* Record the signals used by jvm */
-    jvmsigs |= MASK(sig);
+    /* Record the signals used by jvm. */
+    sigaddset(&jvmsigs, sig);
 
     signal_unlock();
     return res;
@@ -212,9 +299,10 @@ int sigaction(int sig, const struct sigaction *act, struct sigaction *oact) {
   }
 }
 
-/* The three functions for the jvm to call into */
+/* The three functions for the jvm to call into. */
 void JVM_begin_signal_setting() {
   signal_lock();
+  sigemptyset(&jvmsigs);
   jvm_signal_installing = true;
   tid = pthread_self();
   signal_unlock();
@@ -229,8 +317,9 @@ void JVM_end_signal_setting() {
 }
 
 struct sigaction *JVM_get_signal_action(int sig) {
+  allocate_sact();
   /* Does race condition make sense here? */
-  if ((MASK(sig) & jvmsigs) != 0) {
+  if (sigismember(&jvmsigs, sig)) {
     return &sact[sig];
   }
   return NULL;
